refactor: move `TimetableSnapshotParameters` to `framework.transaction` package
