+++ application/src/main/java/org/opentripplanner/framework/transaction/TimetableSnapshotParameters.java
@@ -1,10 +1,16 @@
-package org.opentripplanner.updater;
+package org.opentripplanner.framework.transaction;
 
 import java.time.Duration;
 
 /**
- * {@link
- * org.opentripplanner.standalone.config.routerconfig.UpdatersConfig#timetableUpdates(org.opentripplanner.standalone.config.framework.json.NodeAdapter)}
+ * Parameters governing the behaviour of the timetable snapshot.
+ * <p>
+ * This class lives in the {@code transaction} package even though only {@code maxSnapshotFrequency}
+ * is directly related to transaction concerns, while {@code purgeExpiredData} belongs to the
+ * updater layer. The two parameters cannot be split into separate configuration classes because
+ * doing so would break backwards compatibility of the configuration format. Placing both parameters
+ * here is the only way to avoid circular dependencies between the {@code transaction} and
+ * {@code updater} packages.
  */
 public record TimetableSnapshotParameters(Duration maxSnapshotFrequency, boolean purgeExpiredData) {
   public static final TimetableSnapshotParameters DEFAULT = new TimetableSnapshotParameters(
